+++ src/hooks/use-map-with-points/use-map-with-points.const.ts
@@ -7,6 +7,11 @@ type MarkerMapType = {
   Anchor: [number, number];
 }
 
+export const TileLayerSetting = {
+  Url: 'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
+  Attribution: '&copy; <a href="https://www.openstreetmap.org/copyright" target="_blank">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions" target="_blank">CARTO</a>'
+};
+
 export const MarkerMap: MarkerMapType = {
   Default: '/img/pin.svg',
   Current: '/img/pin-active.svg',
