--- src/pages/login-screen/login-screen.type.ts
@@ -1,4 +1,0 @@
-export type LoginFormType = {
-  email: string;
-  password: string;
-};
